Fixed input height and width sizes

+++ motor.h
@@ -9,7 +9,7 @@
 class Motor {
 public:
   Motor();
-  void setup(Ui::MainWindow *, uint8_t, uint8_t);
+  void setup(Ui::MainWindow *, uint16_t, uint16_t);
   void on(warnings warning);
   void off(warnings warning);
   ~Motor(){};
